refactor(provider): centralize model list logic and improve UI
