+++ NSChat/Views/Provider/ProviderView+Add.swift
@@ -1,47 +1,29 @@
-import SwiftUI
-import SwiftData
 import os
+import SwiftData
+import SwiftUI
 
 struct AddProviderView: View {
   @Environment(\.modelContext) private var modelContext
   @Environment(\.dismiss) private var dismiss
   
-  @State private var selectedType: ProviderType = .openAI
-  @State private var alias: String = ""
-  @State private var apiKey: String = ""
-  @State private var endpoint: String = ""
-  @State private var enabled: Bool = true
-  @State private var fetchedModels: [ModelInfo] = []
-  @State private var fetchStatus: ProviderFetchStatus = .idle
-  
-  private var displayName: String {
-    if !alias.isEmpty {
-      return alias
-    }
-    return selectedType.displayName
-  }
-  
+  @Bindable var provider: Provider = .init(type: .openAI)
+  @State private var searchText = ""
+
   var body: some View {
     NavigationStack {
       List {
         ProviderConfigurationForm(
-          providerType: $selectedType,
-          alias: $alias,
-          apiKey: $apiKey,
-          endpoint: $endpoint,
-          enabled: $enabled
+          providerType: $provider.type,
+          alias: $provider.alias,
+          apiKey: $provider.apiKey,
+          endpoint: $provider.endpoint,
+          enabled: $provider.enabled
         )
         
-        ModelListSection(
-          providerType: selectedType,
-          apiKey: apiKey,
-          endpoint: endpoint,
-          fetchedModels: $fetchedModels,
-          fetchStatus: $fetchStatus
-        )
+        ModelListSection(provider: provider,searchText: $searchText)
       }
-      .navigationTitle("Add Provider")
       .navigationBarTitleDisplayMode(.inline)
+      .searchable(text: $searchText)
       .toolbar {
         ToolbarItem(placement: .cancellationAction) {
           Button("Cancel") {
@@ -53,35 +35,16 @@ struct AddProviderView: View {
           Button("Save") {
             saveProvider()
           }
-          .disabled(apiKey.isEmpty)
+          .disabled(provider.apiKey.isEmpty)
         }
       }
+      .navigationTitle("Add Provider")
     }
   }
   
   private func saveProvider() {
-    let provider = Provider(
-      type: selectedType,
-      alias: alias,
-      apiKey: apiKey,
-      endpoint: endpoint,
-      enabled: enabled
-    )
-    
     modelContext.insert(provider)
-    
-    for modelInfo in fetchedModels {
-      let model = ModelEntity(
-        provider: provider,
-        modelId: modelInfo.id,
-        modelName: modelInfo.name,
-        contextLength: modelInfo.contextLength
-      )
-      modelContext.insert(model)
-    }
-    
-    AppLogger.data.info("Added new provider: \(provider.displayName) with \(fetchedModels.count) models")
-    
+    AppLogger.data.info("Added new provider: \(provider.displayName) with \(provider.models.count) models")
     dismiss()
   }
 }
@@ -91,9 +54,7 @@ struct AddProviderView: View {
     .modelContainer(ModelContainer.preview())
 }
 
-
 #Preview {
   AddProviderView()
     .modelContainer(ModelContainer.preview())
 }
-
